DT000

+++ test/unittest/fms_form_util_test/fms_form_util_test.cpp
@@ -212,7 +212,7 @@ HWTEST_F(FormUtilTest, FormUtilTest_013, TestSize.Level1)
     auto result = FormUtil::StringSplit(str, Constants::TIME_DELIMETER);
 
     EXPECT_TRUE(result.size() == count);
-    for (int64_t i = 0; i < 52; ++i) {
+    for (int i = 0; i < 52; ++i) {
         EXPECT_EQ(FormUtil::TestFormId(i), i);
     }
 }
